refactor: daily components 폴더명 s 제거
