+++ src/app/daily/_components/Streak/Streak.styled.tsx
